Moved NodeType out of NodeVariant
